Move compute_relative_values/1 to Core.Memory
